refactored testutils to have sample types and samples objects by keepers seperated
