+++ x/dao/types/message_mint_token_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-	"github.com/planetmint/planetmint-go/testutil/sample"
+	"github.com/planetmint/planetmint-go/testutil/errormsg"
 	"github.com/stretchr/testify/require"
 )
 
@@ -18,7 +18,7 @@ func TestMsgMintToken_ValidateBasic(t *testing.T) {
 		{
 			name: "invalid address",
 			msg: MsgMintToken{
-				Creator: sample.ErrorInvalidAddress,
+				Creator: errormsg.ErrorInvalidAddress,
 			},
 			err: sdkerrors.ErrInvalidAddress,
 		},
